move timezones to module
